refactor: 스케줄러 통합
